requirements.txt
tornado
supervisor
wifi
psutil
pillow
pygame
RPi.GPIO

